// src/app/shared/generated/model/agentFilter.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AgentFilter { 
    key?: AgentFilterKeyEnum;
    value?: string;
}
export enum AgentFilterKeyEnum {
    AppId = 'APP_ID'
};