// src/app/shared/generated/model/agentGroup.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AgentGroupResponseStrategy } from './agentGroupResponseStrategy';
import { AgentGroupOrchestrationMode } from './agentGroupOrchestrationMode';


export interface AgentGroup { 
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    name?: string;
    description?: string;
    routingInstructions?: string;
    orchestrationMode?: AgentGroupOrchestrationMode;
    responseStrategy?: AgentGroupResponseStrategy;
}