// src/app/shared/generated/model/model.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CommunicationMode } from './communicationMode';
import { Provider } from './provider';


export interface Model { 
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    name?: string;
    modelIdentifier?: string;
    modelConfig?: string;
    communicationMode?: CommunicationMode;
    provider?: Provider;
}