// src/app/shared/generated/model/scaffold.ts
/**
 * onecx-ai bff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Skill } from './skill';


export interface Scaffold { 
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    name?: string;
    systemPrompt?: string;
    sourceProduct?: string;
    skills?: Array<Skill>;
}