// src/app/shared/generated/model/tool.ts
/**
 * onecx-ai-provider  service
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ExecutionPolicy } from './executionPolicy';
import { ToolType } from './toolType';
import { AuthMode } from './authMode';


export interface Tool { 
    modificationCount?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    name?: string;
    description?: string;
    type?: ToolType;
    source?: ToolSourceEnum;
    url?: string;
    apiKey?: string;
    executionPolicy?: ExecutionPolicy;
    authMode?: AuthMode;
}
export enum ToolSourceEnum {
    Tenant = 'TENANT',
    Global = 'GLOBAL'
};